requirements.txt
# Object Detection & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
ultralytics>=8.0.0
pillow>=10.0.0
numpy>=1.24.0

# Utilities
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0

# Web Interface
streamlit>=1.28.0

# Web API (opcional)
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
